requirements.txt
python-telegram-bot>=21.6,<22
anthropic>=0.40.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.36,<3
python-dotenv>=1.0.0
httpx>=0.28.0
